+++ src/index.ts
@@ -1,3 +1,7 @@
+import * as constants from './constants';
+
 export * from './configure';
 export * from './types';
 export * from './service';
+
+export { constants };